+++ src/modules/Elsa.ProtoActor.Core/Abstractions/VirtualActorsProviderBase.cs
@@ -1,0 +1,11 @@
+using Google.Protobuf.Reflection;
+using Proto;
+using Proto.Cluster;
+
+namespace Elsa.ProtoActor;
+
+public abstract class VirtualActorsProviderBase : IVirtualActorsProvider
+{
+    public abstract IEnumerable<ClusterKind> GetClusterKinds(ActorSystem system);
+    public virtual IEnumerable<FileDescriptor> GetFileDescriptors() => [];
+}
