+++ src/modules/Elsa.ProtoActor.Core/Elsa.ProtoActor.Core.csproj
@@ -1,0 +1,29 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <Description>
+            Provides ProtoActor to the system. Other modules can add actors to the ActorSystem.
+        </Description>
+        <PackageTags>elsa module runtime protoactor actors</PackageTags>
+        <RootNamespace>Elsa.ProtoActor</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions"/>
+    </ItemGroup>
+
+    <ItemGroup Label="ProtoActor">
+        <PackageReference Include="Proto.Actor"/>
+        <PackageReference Include="Proto.Cluster"/>
+        <PackageReference Include="Proto.Cluster.TestProvider"/>
+        <PackageReference Include="Proto.OpenTelemetry"/>
+        <PackageReference Include="Proto.Persistence"/>
+        <PackageReference Include="Proto.Remote"/>
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\common\Elsa.Features\Elsa.Features.csproj" />
+      <ProjectReference Include="..\Elsa.Workflows.Core\Elsa.Workflows.Core.csproj" />
+    </ItemGroup>
+    
+</Project>
