+++ src/modules/Elsa.ProtoActor.Core/Extensions/WorkflowRuntimeFeatureExtensions.cs
@@ -1,0 +1,17 @@
+using Elsa.Features.Services;
+using Elsa.ProtoActor.Features;
+using JetBrains.Annotations;
+
+// ReSharper disable once CheckNamespace
+namespace Elsa.Extensions;
+
+[PublicAPI]
+public static class ProtoActorModuleExtensions
+{
+    /// Creates and configures a new ActorSystem. 
+    public static IModule UseProtoActor(this IModule module, Action<ProtoActorFeature>? configure = default)
+    {
+        module.Configure(configure);
+        return module;
+    }
+}
