+++ src/modules/Elsa.ProtoActor.Core/HostedServices/StartClusterMember.cs
@@ -1,0 +1,15 @@
+using Microsoft.Extensions.Hosting;
+using Proto;
+using Proto.Cluster;
+
+namespace Elsa.ProtoActor.HostedServices;
+
+/// Starts the current member in the cluster.
+public class StartClusterMember(ActorSystem actorSystem) : IHostedService
+{
+    /// Starts the current member in the cluster.
+    public Task StartAsync(CancellationToken cancellationToken) => actorSystem.Cluster().StartMemberAsync();
+
+    /// Stops the current member in the cluster.
+    public Task StopAsync(CancellationToken cancellationToken) => actorSystem.Cluster().ShutdownAsync();
+}
