+++ src/modules/Elsa.Workflows.Runtime.ProtoActor/Providers/WorkflowInstanceVirtualActorProvider.cs
@@ -1,0 +1,14 @@
+// using Elsa.ProtoActor;
+// using Elsa.Workflows.Runtime.ProtoActor.Actors;
+// using Proto;
+// using Proto.Cluster;
+//
+// namespace Elsa.Workflows.Runtime.ProtoActor.Providers;
+//
+// public class WorkflowInstanceVirtualActorProvider  : VirtualActorProviderBase
+// {
+//     public override ClusterKind CreateClusterKind(ActorSystem system)
+//     {
+//         new ClusterKind(WorkflowInstanceImpl)
+//     }
+// }
